Added tweaks like shadows, messing around with tooltip, and also start on relay, host is done!

+++ survival-project/Assets/Scripts/ItemScripts/InventoryItemData.cs
@@ -25,7 +25,7 @@ public class InventoryItemData : ScriptableObject
     public GameObject ItemPrefab;
 
     [Header("Weapons and Tools")]
-    public double itemDamage;
+    public float itemDamage;
     public GameObject projectilePrefab;
     public float projectileSpeed;
     public float projectileLifetime;
